Add custom actor RPM packaging and build instructions (#2)
* Allow packaging custom actors into RPM and provide build instructions

Inspired from leapp-repository spec file and makefile

The specfile is designed to package the custom actors into an RPM,
and the actor-list.txt file allows maintainers to add or remove actors
that should be included in the RPM.

Depending on the DIST_VERSION env var, either RHEL7 or RHEL8 upgrade
repositories are created under
`/usr/share/leapp-repositories/repositories/system_upgrade_supplements`
- Run make rpmbuild to install the default RHEL7 supplement repos.
- Run make rpmbuild DIST_VERSION=8 for RHEL8 version.

Customers can choose which actors to install. This is allowed by editing
the `actor-list.txt` file. Each actor is defined on its own line and is
parsed during the rpmbuild process. Comments are defined by using `#`.

`py_byte_compile` is used to compile the actors during install.

The README.md file has been updated with a new section that provides
detailed build instructions and prerequisites for users who want to
package their custom actors into an RPM.

* Change license from MIT to Apache 2.0

This change is made to be consistent with leapp-repository

* Update README.md

---------

Co-authored-by: Marek Filip <mafilip@redhat.com>
Co-authored-by: Bob Mader <bmader@redhat.com>